+++ src/App.js
@@ -1,91 +1,18 @@
-import React, { useState } from "react";
-import SubmitButton from "./components/SubmitButton"
-import SignupButton from "./components/SignupButton";
-
-import "./styles.css";
+import React from "react";
+import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
+import Login from "./Pages/Login";
+import Homepage from "./Pages/Homepage";
 
 function App() {
-  // React States
-  const [errorMessages, setErrorMessages] = useState({});
-  const [isSubmitted, setIsSubmitted] = useState(false);
+  return(
+    <Router>
+      <Switch>
+        <Route exact path="/login" component = {Login}/>
+        <Route exact path="/homepage" component = {()=><Homepage authorized={true}/>}/>
+      </Switch>
+    </Router>
+  )
 
-  // User Login info
-  const database = [
-    {
-      username: "name",
-      password: "password"
-    },
-    {
-      username: "name1",
-      password: "password1"
-    }
-  ];
-
-  const errors = {
-    uname: "invalid username",
-    pass: "invalid password"
-  };
-
-  const handleSubmit = (event) => {
-    //Prevent page reload
-    event.preventDefault();
-
-    var { uname, pass } = document.forms[0];
-
-    // Find user login info
-    const userData = database.find((user) => user.username === uname.value);
-
-    // Compare user info
-    if (userData) {
-      if (userData.password !== pass.value) {
-        // Invalid password
-        setErrorMessages({ name: "pass", message: errors.pass });
-      } else {
-        setIsSubmitted(true);
-      }
-    } else {
-      // Username not found
-      setErrorMessages({ name: "uname", message: errors.uname });
-    }
-  };
-
-  // Generate JSX code for error message
-  const renderErrorMessage = (name) =>
-    name === errorMessages.name && (
-      <div className="error">{errorMessages.message}</div>
-    );
-
-  // JSX code for login form
-  const renderForm = (
-    <div className="form">
-      <form onSubmit={handleSubmit}>
-        <div className="input-container">
-          <label>Username </label>
-          <input type="text" name="uname" required />
-          {renderErrorMessage("uname")}
-        </div>
-        <div className="input-container">
-          <label>Password </label>
-          <input type="password" name="pass" required />
-          {renderErrorMessage("pass")}
-        </div>
-        <SubmitButton/>
-        <span class="forgotPassword">
-          <a href="aaa.html">Forgot Password?</a>
-        </span>
-        <SignupButton style={{textAlign: 'center', top:'100px'}}/>
-      </form>
-    </div>
-  );
-
-  return (
-    <div className="app">
-      <div className="login-form">
-        <div className="title" style={{textAlign: 'center', marginBottom:'30px'}}>Smart Placement</div>
-        {isSubmitted ? <div>User is successfully logged in</div> : renderForm}
-      </div>
-    </div>
-  );
 }
 
 export default App;
