+++ src/Style/homepage.css
@@ -1,0 +1,152 @@
+*{
+    margin: 0;
+    padding: 0;
+    font-family: sans-serif;
+}
+
+.banner{
+    width: 100%;
+    height: 100vh;
+    background-image: linear-gradient(rgba(0,0,0,0.2), rgba(0,0,0,0.2));
+    background-color: rgb(248, 248, 248);
+    background-size: cover;
+    background-position: center;
+}
+
+.navbar{
+    width: 85%;
+    margin: auto;
+    padding: 0;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.logo{
+    width: 120px;
+    cursor: pointer;
+}
+
+.navbar ul li {
+    list-style: none;
+    display: inline-block;
+    margin: 0 20px;
+    position: relative;
+}
+
+.navbar ul li a{
+    text-decoration: none;
+    color: purple;
+    font-weight: bold;
+    text-transform: uppercase;
+}
+
+.navbar ul li a:hover{
+    color: blue;
+    border-radius: 20px;
+    border: 2px solid rgb(0, 255, 200);
+    padding: 8px;
+}
+
+.navbar ul li::after {
+    content: '';
+    height: 2px;
+    width: 0;
+    background: orange;
+    position: absolute;
+    left: 0;
+    bottom: -15px;
+    transition: 0.5s;
+}
+
+.navbar ul li:hover::after {
+    width: 100%;
+    border-radius: 15px;
+    border: 2px solid red;
+}
+
+.content{
+    width: 100%;
+    position: absolute;
+    top: 50%;
+    transform: translateY(-50%);
+    text-align: center;
+    color: blue;
+}
+
+.content h2{
+    font-size: 50px;
+    margin-top: 60px;
+}
+
+.content p{
+    margin: 20px auto;
+    font-weight: 100;
+    line-height: 25px;
+    color: rgb(199, 7, 247);
+}
+
+.newButton{
+    width: 200px;
+    padding: 15px 0;
+    text-align: center;
+    margin: 20px 10px;
+    border-radius: 25px;
+    font-weight: bold;
+    border: 2px solid orange;
+    background: transparent;
+    color: purple;
+    cursor: pointer;
+    position: relative;
+    overflow: hidden;
+}
+
+span{
+    background: rgb(3, 248, 215);
+    height: 100%;
+    width: 0;
+    border-radius: 25px;
+    position: absolute;
+    left: 0;
+    bottom: 0;
+    z-index: -1;
+    transition: 0.5s;
+}
+
+.newButton:hover span{
+    width: 100%;
+}
+
+.newButton:hover{
+    border: none;
+    text-transform: uppercase;
+    color: blue;
+}
+
+.logout{
+    text-align: right;
+    background-color: black;
+}
+
+.logout a{
+    border-radius: 8px;
+    text-align: center;
+    border: 2px solid rgb(0, 255, 200);
+    padding: 10px 25px;
+    text-decoration: none;
+    text-transform: uppercase;
+    font-size: 14px;
+    margin: 20px 10px;
+    display: inline-block;
+    color: purple;
+    width: 8%;
+    background-color: gold;
+    font-weight: bold;
+}
+    
+.logout a:hover {
+    background: red;
+    color: blue;
+    transition: .5s;
+    font-weight: bold;
+}
