+++ src/components/SignupButton.js
@@ -2,8 +2,8 @@ import React from 'react'
 
 const SignupButton = ({style}) => {
   return (
-    <div className="signupButton" style={style}>
-        <button>Sign Up</button>
+    <div style={style}>
+        <button className="signupButton" >Sign Up</button>
     </div>
   )
 }
